[JavaWeaver] Adds EAPK library
Energy-aware Android Patterns is a library to look for energy-inefficient patterns in android source code.
This library works in tandem with the 'Compile with incomplete classpath' and JadxPlus
to automatically decompile APK into source code.
This allows to not only analyse working projects, but also works with code fragments and decompiled APKs.